move locobot's perception into a robot/ folder temporarily
[ghstack-poisoned]
