+++ droidlet/memory/robot/tests/test_memory.py
@@ -5,10 +5,10 @@ import unittest
 import numpy as np
 from numpy.testing import assert_allclose
 from droidlet.lowlevel.locobot.locobot_mover_utils import xyz_canonical_coords_to_pyrobot_coords
-from droidlet.perception import Detection, RGBDepth
+from droidlet.perception.robot import Detection, RGBDepth
 from droidlet.memory.robot.loco_memory import LocoAgentMemory
 from droidlet.memory.robot.loco_memory_nodes import DetectedObjectNode, HumanPoseNode
-from droidlet.perception.tests.utils import get_fake_detection, get_fake_humanpose, get_fake_bbox
+from droidlet.perception.robot.tests.utils import get_fake_detection, get_fake_humanpose, get_fake_bbox
 
 
 class MemoryTests(unittest.TestCase):
