--- droidlet/perception/__init__.py
@@ -1,42 +1,0 @@
-from .perception_helpers import (
-    random_colors,
-    draw_xyz,
-    get_random_color,
-    get_coords,
-    get_color_tag,
-)
-from .handlers import (
-    InputHandler,
-    MemoryHandler,
-    DetectionHandler,
-    FaceRecognitionHandler,
-    ObjectDeduplicationHandler,
-    Detection,
-    WorldObject,
-    RGBDepth,
-    Human,
-    HumanKeypoints,
-)
-from .perception import Perception, SlowPerception
-from .self_perception import SelfPerception
-
-__all__ = [
-    Perception,
-    SlowPerception,
-    SelfPerception,
-    Detection,
-    WorldObject,
-    Human,
-    HumanKeypoints,
-    RGBDepth,
-    random_colors,
-    draw_xyz,
-    get_random_color,
-    get_coords,
-    get_color_tag,
-    InputHandler,
-    DetectionHandler,
-    MemoryHandler,
-    FaceRecognitionHandler,
-    ObjectDeduplicationHandler,
-]
