+++ droidlet/perception/robot/__init__.py
@@ -1,0 +1,42 @@
+from droidlet.perception.robot.perception_helpers import (
+    random_colors,
+    draw_xyz,
+    get_random_color,
+    get_coords,
+    get_color_tag,
+)
+from droidlet.perception.robot.handlers import (
+    InputHandler,
+    MemoryHandler,
+    DetectionHandler,
+    FaceRecognitionHandler,
+    ObjectDeduplicationHandler,
+    Detection,
+    WorldObject,
+    RGBDepth,
+    Human,
+    HumanKeypoints,
+)
+from droidlet.perception.robot.perception import Perception, SlowPerception
+from droidlet.perception.robot.self_perception import SelfPerception
+
+__all__ = [
+    Perception,
+    SlowPerception,
+    SelfPerception,
+    Detection,
+    WorldObject,
+    Human,
+    HumanKeypoints,
+    RGBDepth,
+    random_colors,
+    draw_xyz,
+    get_random_color,
+    get_coords,
+    get_color_tag,
+    InputHandler,
+    DetectionHandler,
+    MemoryHandler,
+    FaceRecognitionHandler,
+    ObjectDeduplicationHandler,
+]
