+++ droidlet/perception/robot/detectron/configs/Keypoints/Base-RCNN-FPN.yaml
