+++ droidlet/perception/robot/detectron/configs/Keypoints/COCO-Keypoints/Base-Keypoint-RCNN-FPN.yaml
