+++ droidlet/perception/robot/detectron/configs/Keypoints/COCO-Keypoints/keypoint_rcnn_R_50_FPN_3x.yaml
