+++ droidlet/perception/robot/detectron/detector/README.MD
