+++ droidlet/perception/robot/detectron/detector/configs/Base-RCNN-FPN.yaml
