+++ droidlet/perception/robot/detectron/detector/configs/mask_rcnn_R_101_FPN_1x.yaml
