+++ droidlet/perception/robot/detectron/detector/properties_head.py
