+++ droidlet/perception/robot/handlers/laser_pointer.py
