+++ droidlet/perception/robot/offline_files/faces/Mirghany.jpg
