+++ droidlet/perception/robot/offline_files/faces/arthur_szlam.png
