+++ droidlet/perception/robot/offline_files/faces/bill_gates.jpg
