+++ droidlet/perception/robot/offline_files/faces/jeff bezos.jpg
