+++ droidlet/perception/robot/offline_files/faces/obama.jpg
