+++ droidlet/perception/robot/offline_files/faces/sara.jpg
