+++ droidlet/perception/robot/offline_files/faces/stephan.jpg
