+++ droidlet/perception/robot/offline_files/videos/chair_only.mp4
