+++ droidlet/perception/robot/tests/utils.py
@@ -6,7 +6,7 @@ import numpy as np
 from droidlet.dialog.dialogue_manager import DialogueManager
 from droidlet.dialog.droidlet_nsp_model_wrapper import DroidletNSPModelWrapper
 from agents.loco_mc_agent import LocoMCAgent
-from droidlet.perception import Detection, RGBDepth, Human, HumanKeypoints
+from droidlet.perception.robot import Detection, RGBDepth, Human, HumanKeypoints
 
 
 class FakeAgent(LocoMCAgent):
